requirements.txt
amqp==2.2.1
billiard==3.5.0.3
celery==4.1.0
chaussette==1.3.0
circus==0.14.0
diff-match-patch==20121119
Django==1.11.3
django-celery-beat==1.0.1
django-celery-results==1.0.1
django-import-export==0.5.1
django-material==1.0.0
et-xmlfile==1.0.1
iowait==0.2
jdcal==1.3
kombu==4.1.0
odfpy==1.3.5
openpyxl==2.4.8
psutil==5.2.2
psycopg2-binary==2.7.4
pytz==2017.2
PyYAML==3.12
pyzmq==16.0.2
six==1.10.0
tablib==0.12.1
tornado==4.5.1
unicodecsv==0.14.1
vine==1.1.4
xlrd==1.1.0
xlwt==1.3.0
